--- examples/llm/tape/tape/dataset/utils.py
@@ -1,8 +1,0 @@
-import hashlib
-
-
-def generate_string_hash(input_string: str, algorithm: str = 'sha256'):
-    input_bytes = input_string.encode('utf-8')
-    hash_obj = hashlib.new(algorithm)
-    hash_obj.update(input_bytes)
-    return hash_obj.hexdigest()
